update dependency testing
Drop 10.14 from the testing suite as anything older than 10.26.x
conflicts with the testing tools, and it's no longer a supported
verson from the maintainer. We will continue to test when possible.

Add lock files for appraisal gems to prevent version mishaps when
newer versions are produced.

Tell Travis to test these versions.