:lipstick: enable Datadog event alerts on Chef failure
When a given Chef run_status is anything other than `success`, the
Event body will add any @ handles that have been provided in an array
through the Handler's configuration.

This feature will go unused until the chef-datadog cookbook has been
updated to use it, which is contingent on determining if there is a
sane upgrade path that does not trigger upgrade of the handler
unintentionally.

Fixes #29