move tag collection to earlier in the event method
This was initially placed in the tag updater method, since historically
we would retrieve tags from Datadog prior to updating, making the
existence of an application_key a prerequisite.

Moving this local-only method to higher in the method exposes the tags
variables to be used in the Event method as well.

Refs #52